review

+++ src/main/java/life/toughjourney/community/controller/QuestionController.java
@@ -1,15 +1,12 @@
 package life.toughjourney.community.controller;
 
 import life.toughjourney.community.dto.QuestionDto;
-import life.toughjourney.community.mapper.QuestionMapper;
 import life.toughjourney.community.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-
-import javax.jws.WebParam;
 
 /**
  * @author YX
@@ -23,7 +20,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("question/{id}")
-    public String question(@PathVariable(name = "id")Integer id,
+    public String question(@PathVariable(name = "id")Long id,
                            Model model){
         QuestionDto questionDto = questionService.getById(id);
         //累加阅读数
